+++ src/components/ApplicationViews.js
@@ -13,7 +13,7 @@ import { PostsByUser } from "./posts/PostsByUser.js"
 import { SinglePost } from "./posts/SinglePost.js"
 import { EditCategoryForm } from "./categories/EditCategory.js"
 import { UpdateTagForm } from "./tags/UpdateTags.js"
-import { ApprovePosts } from "./posts/ApprovePosts.js"
+import {  ApprovePostsList } from "./posts/ApprovePosts.js"
 
 export const ApplicationViews = () => {
   return (
@@ -49,7 +49,7 @@ export const ApplicationViews = () => {
         <MyPosts />
       </Route>
       <Route exact path="/posts/approve">
-        <ApprovePosts />
+        <ApprovePostsList />
       </Route>
       <Route exact path="/posts/user/:userId(\d+)">
         <PostsByUser />
