+++ src/components/posts/Post.js
@@ -6,7 +6,7 @@ import { ButtonControls } from "../buttonControls/ButtonControls"
 import { CommentList } from "../comments/CommentsList"
 import { TagsList } from "../tags/TagsList"
 import "./Post.css"
-import { getSinglePost } from "./PostManager"
+import { getSinglePost,editPost } from "./PostManager"
 
 
 // function that renders a single post
@@ -48,6 +48,21 @@ export const Post = ({ listView, cardView, post, setPost }) => {
             })
         }},[postRefresh]
     )
+
+    const ApprovePostClick = (evt) => {
+        evt.preventDefault()
+
+        const editingPostForApproval = {
+            id: parseInt(selectPost.id),
+            category: parseInt(selectPost.category.id),
+            publication_date: selectPost.publication_date,
+            title: selectPost.title,
+            image_url: selectPost.image_url,
+            content: selectPost.content,
+            approved: 1
+        }
+        editPost(editingPostForApproval).then(()=>history.push(`/posts/approve`))
+    }
 
     return <>
         {/* Content needed in all posts list */}
@@ -135,6 +150,11 @@ export const Post = ({ listView, cardView, post, setPost }) => {
                                 ? <button onClick={() => { setShowTagBoxes(!showTagBoxes) }}>Manage Tags</button>
                                 : null
                                 
+                            }
+                            {
+                                post.approved === false 
+                                ? <button onClick={ApprovePostClick}>Approve</button>
+                                : null
                             }
                             </div>
                         </div>
